move remaining functions from install to archive utils
