+++ localstack/services/dynamodb/packages.py
@@ -4,7 +4,7 @@ from typing import List
 
 from localstack.constants import ARTIFACTS_REPO, MAVEN_REPO_URL
 from localstack.packages import InstallTarget, Package, PackageInstaller
-from localstack.services.install import (
+from localstack.utils.archives import (
     download_and_extract_with_retry,
     update_jar_manifest,
     upgrade_jar_file,
